publish function call good

+++ src/main/resources/liquibase/changelogs/functions/publish.sql
@@ -1,5 +1,5 @@
 
-create or replace function publish(eid numeric, d varchar)
+create or replace function publish(eid integer, d varchar)
 returns void as
 $BODY$
 BEGIN
